remove debris of files and code. re-structure the project and housekeeping
