+++ requirements.txt
@@ -1,5 +1,23 @@
-# for development and testing
-pytest == 6.1.2
-pytest-cov == 2.11.1
-pylint == 2.5.3
-beautifulsoup4 == 4.9.3
+astroid==2.5
+attrs==21.2.0
+beautifulsoup4==4.9.3
+certifi==2021.5.30
+chardet==4.0.0
+coverage==5.5
+idna==2.10
+iniconfig==1.1.1
+isort==4.3.21
+lazy-object-proxy==1.6.0
+mccabe==0.6.1
+packaging==20.9
+pluggy==0.13.1
+py==1.10.0
+pylint==2.5.3
+pyparsing==2.4.7
+pytest==6.1.2
+pytest-cov==2.11.1
+requests==2.25.1
+soupsieve==2.2.1
+toml==0.10.2
+urllib3==1.26.5
+wrapt==1.12.1
