--- scraper-kalina/README.MD
@@ -1,1 +1,0 @@
-Web scraper
