--- scraper-vnts/test_scraper.py
@@ -1,10 +1,0 @@
-import pytest
-import scraper
-
-scrp = scraper.Scraper()
-
-scrp.getPostsURLs()
-scrp.getPosts(scrp.listURLS)
-
-def test_connection():
-	assert scrp.srouce.status_code == 200
