+++ template/.coveragerc
@@ -1,0 +1,28 @@
+# .coveragerc to control coverage.py
+[run]
+branch = True
+cover_pylib = False
+include =
+    main.py
+    module/*
+omit =
+    test/*
+
+[report]
+# Regexes for lines to exclude from consideration
+exclude_lines =
+    # Have to re-enable the standard pragma
+    pragma: no cover
+
+    # Don't complain if tests don't hit defensive assertion code:
+    raise AssertionError
+    raise NotImplementedError
+
+    # Don't complain if non-runnable code isn't run:
+    if 0:
+    if __name__ == .__main__.:
+
+ignore_errors = True
+
+[pytest]
+addopts = --cov-report html --cov-fail-under=90 --cov=.
